+++ src/Table/Columns/Concerns/HasItems.php
@@ -81,8 +81,8 @@ trait HasItems
         } else {
             $result = call_user_func($this->itemsResolver, $this->getModel());
         }
-
-        $result = $result instanceof Collection ? $result : collect($result);
+        //        $result = $result instanceof Collection ? $result : collect($result);
+        $result = $result instanceof Collection ? $result : (! is_iterable($result) ? collect([$result]) : collect($result));
 
         return $result->values()->map(function (mixed $rawItem) {
             return (! $rawItem instanceof ColumnItem) ? $this->replicateToItem($rawItem) : $rawItem;